requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
coverage==7.2.7
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.99.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.10
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
yarl==1.9.2
